Related work items: #309, #310, #311

+++ DijiWalk.Repositories/GameRepository.cs
@@ -9,6 +9,7 @@ namespace DijiWalk.Repositories
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
+    using DijiWalk.Business.Contracts;
     using DijiWalk.Common.Contracts;
     using DijiWalk.Common.Response;
     using DijiWalk.Entities;
@@ -25,14 +26,20 @@ namespace DijiWalk.Repositories
 
         private readonly SmartCityContext _context;
 
+        private readonly IGameBusiness _gameBusiness;
+
+        private readonly IPlayBusiness _playBusiness;
+
+
         /// <summary>
         /// Parameter that serve to connect to the database
         /// </summary>
-        public GameRepository(SmartCityContext context)
+        public GameRepository(SmartCityContext context, IGameBusiness gameBusiness, IPlayBusiness playBusiness)
         {
             this._context = context;
+            this._gameBusiness = gameBusiness;
+            this._playBusiness = playBusiness;
         }
-
         /// <summary>
         /// Method to Add the Game passed in the parameters to the database
         /// </summary>
@@ -43,7 +50,7 @@ namespace DijiWalk.Repositories
             {
                 await _context.Games.AddAsync(game);
                 _context.SaveChanges();
-                return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Add, Response = game };
+                return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Add, Response = await this.Find(game.Id) };
             }
             catch (Exception e)
             {
@@ -59,9 +66,17 @@ namespace DijiWalk.Repositories
         {
             try
             {
-                _context.Games.Remove(await _context.Games.FindAsync(id));
-                _context.SaveChanges();
-                return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Delete };
+                var game = await _context.Games.FindAsync(id);
+                if (_gameBusiness.IsGameInProgress(game) == false)
+                {
+                    await _gameBusiness.DeleteAllTeams(game);
+                    _context.Games.Remove(await _context.Games.FindAsync(id));
+                    _context.SaveChanges();
+                    return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Delete };
+                } else
+                {
+                    return new ApiResponse { Status = ApiStatus.CantDelete, Message = "Ce jeu est en cours.." };
+                }
             }
             catch (Exception e)
             {
@@ -85,7 +100,7 @@ namespace DijiWalk.Repositories
         /// <returns>The Game with the Id researched</returns>
         public async Task<Game> Find(int id)
         {
-            return await _context.Games.FindAsync(id);
+            return await _context.Games.Where(g => g.Id == id).Include(r => r.Route).Include(t => t.Transport).Include(o => o.Organizer).Include(p => p.Plays).ThenInclude(t => t.Team).ThenInclude(m => m.TeamPlayers).ThenInclude(p => p.Player).FirstOrDefaultAsync();
         }
 
         /// <summary>
@@ -94,17 +109,52 @@ namespace DijiWalk.Repositories
         /// <returns>A List with all Games</returns>
         public async Task<IEnumerable<Game>> FindAll()
         {
-            return await _context.Games.ToListAsync();
+            return await _context.Games
+                .Include(r => r.Route)
+                .Include(t => t.Transport)
+                .Include(o => o.Organizer)
+                .Include(p => p.Plays).ThenInclude(t => t.Team)
+                .ToListAsync();
         }
 
         /// <summary>
         /// Method that will Update the Game passed in the parameters to the database
         /// </summary>
         /// <param name="game">Object Game to Update</param>
-        public void Update(Game game)
+        public async Task<ApiResponse> Update(Game game)
         {
-            _context.Games.Update(game);
-            _context.SaveChanges();
+            try
+            {
+                _context.Games.Update(_gameBusiness.SeparatePlay(game));
+                await _context.SaveChangesAsync();
+
+                // Récupérer tous les id actuels des teams de la game dans la DB
+                var currentIdPlays = await _context.Plays.Where(p => p.IdGame == game.Id).Select(p => p.IdTeam).ToListAsync();
+
+                // Récuperer tous les id teams de la game envoyée en paramètre
+                var idPlay = game.Plays.Select(p => p.IdTeam).ToList();
+
+                // liste des id de la BDD qui ne sont pas dans les ID de l'update
+                var oldIdPlay = currentIdPlays.Where(p => !idPlay.Contains(p)).ToList();
+
+                var responseDelete = await _playBusiness.DeleteFromGame(await _context.Plays.Where(p => p.IdGame == game.Id && oldIdPlay.Contains(p.IdTeam)).ToListAsync());
+                if (responseDelete.Status == ApiStatus.Ok)
+                {
+                    var responsePlayAdd = await _playBusiness.SetUp(game.Plays.Where(p => !currentIdPlays.Contains(p.IdTeam)).ToList(), game.Id);
+                    if (responsePlayAdd.Status == ApiStatus.Ok)
+                    {
+                        return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Update, Response = await this.Find(game.Id) };
+                    }
+                    else
+                        return responsePlayAdd;
+                }
+                else
+                    return responseDelete;
+            }
+            catch (Exception e)
+            {
+                throw;
+            }
         }
     }
 }
